# gym-backend-api/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather API",
    "description": "JSON REST API for viewing weather data"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/class": {
      "post": {
        "summary": "Creates a class",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Creates a class",
          "content": {
            "application/json": {
              "schema": {
                "class_activity_id": "string",
                "class_room_id": "string",
                "class_trainer_user_id": "string",
                "class_date": "string",
                "class_time": "string"
              },
              "example": {
                "class_activity_id": "1",
                "class_room_id": "1",
                "class_trainer_user_id": "2",
                "class_date": "2023-04-04",
                "class_time": "00:00:00"
              }
            }
          }
        }
      }
    },
    "/classes": {
      "get": {
        "summary": "Get all classes",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/class/{id}": {
      "get": {
        "summary": "Get a class by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Class id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "summary": "Update a class by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update a class by id",
          "content": {
            "application/json": {
              "schema": {
                "class_id": "string",
                "class_activity_id": "string",
                "class_room_id": "string",
                "class_trainer_user_id": "string",
                "class_date": "string",
                "class_time": "string"
              },
              "example": {
                "class_id": "1",
                "class_activity_id": "1",
                "class_room_id": "1",
                "class_trainer_user_id": "1",
                "class_date": "2023-04-04",
                "class_time": "00:00:00"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a class by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Class id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/classes/details": {
      "get": {
        "summary": "Get all classes with details",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a class with details",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Create a class with their details",
          "content": {
            "application/json": {
              "schema": {
                "class_activity_name": "Yoga",
                "class_room_number": "1",
                "class_trainer_first_name": "John",
                "class_date": "2021-05-01",
                "class_time": "10:00:00"
              },
              "example": {
                "class_activity_name": "Yoga",
                "class_room_number": "1",
                "class_trainer_first_name": "Arnold",
                "class_date": "2023-05-01",
                "class_time": "10:00:00"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a class with details",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update a class with their details",
          "content": {
            "application/json": {
              "schema": {
                "class_id": "string",
                "class_activity_name": "string",
                "class_room_number": "string",
                "class_trainer_first_name": "string",
                "class_date": "string",
                "class_time": "string"
              },
              "example": {
                "class_id": "1",
                "class_activity_name": "Yoga",
                "class_room_number": "1",
                "class_trainer_first_name": "Chuck",
                "class_date": "2021-05-01",
                "class_time": "10:00:00"
              }
            }
          }
        }
      }
    },
    "/classes/details/{id}": {
      "get": {
        "summary": "Get class by id with details",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Class ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/classes/upload/xml": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}